Code for getting relation types.
This should hopefully be pretty bullet-proof. The code loads (on
demand) editing pages and looks at the numerical values that the
various relationship types have been given.

Unfortunately, there's not an obvious mapping from the web service's
"type" (for example "performance" or "instrument") to the text in the
dropdown box. So this has to be done manually at the moment, I'm
afraid.